Consolidate all the Videx VideoTerm code
