+++ pico/CMakeLists.txt
@@ -29,7 +29,6 @@ target_sources(applevga PUBLIC
     main.c
     abus.c
     buffers.c
-    card_videx.c
     colors.c
     config.c
     device_regs.c
@@ -38,9 +37,9 @@ target_sources(applevga PUBLIC
     render_lores.c
     render_testpat.c
     render_text.c
-    render_videx_text.c
     ${TEXTFONT_SRC_FILES}
     vga.c
+    videx_vterm.c
 )
 target_include_directories(applevga PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 target_link_libraries(applevga PUBLIC
