--- pico/card_videx.c
@@ -1,216 +1,0 @@
-#include "card_videx.h"
-
-
-const uint8_t videx_rom[0x400] = {  // 0x0400 at C800-CBFF
-    0xad, 0x7b, 0x07, 0x29, 0xf8, 0xc9, 0x30, 0xf0, 0x21, 0xa9, 0x30, 0x8d, 0x7b, 0x07, 0x8d, 0xfb, 0x07, 0xa9, 0x00, 0x8d, 0xfb, 0x06,
-    0x20, 0x61, 0xc9, 0xa2, 0x00, 0x8a, 0x8d, 0xb0, 0xc0, 0xbd, 0xa1, 0xc8, 0x8d, 0xb1, 0xc0, 0xe8, 0xe0, 0x10, 0xd0, 0xf1, 0x8d, 0x59,
-    0xc0, 0x60, 0xad, 0xfb, 0x07, 0x29, 0x08, 0xf0, 0x09, 0x20, 0x93, 0xfe, 0x20, 0x22, 0xfc, 0x20, 0x89, 0xfe, 0x68, 0xa8, 0x68, 0xaa,
-    0x68, 0x60, 0x20, 0xd1, 0xc8, 0xe6, 0x4e, 0xd0, 0x02, 0xe6, 0x4f, 0xad, 0x00, 0xc0, 0x10, 0xf5, 0x20, 0x5c, 0xc8, 0x90, 0xf0, 0x2c,
-    0x10, 0xc0, 0x18, 0x60, 0xc9, 0x8b, 0xd0, 0x02, 0xa9, 0xdb, 0xc9, 0x81, 0xd0, 0x0a, 0xad, 0xfb, 0x07, 0x49, 0x40, 0x8d, 0xfb, 0x07,
-    0xb0, 0xe7, 0x48, 0xad, 0xfb, 0x07, 0x0a, 0x0a, 0x68, 0x90, 0x1f, 0xc9, 0xb0, 0x90, 0x1b, 0x2c, 0x63, 0xc0, 0x30, 0x14, 0xc9, 0xb0,
-    0xf0, 0x0e, 0xc9, 0xc0, 0xd0, 0x02, 0xa9, 0xd0, 0xc9, 0xdb, 0x90, 0x08, 0x29, 0xcf, 0xd0, 0x04, 0xa9, 0xdd, 0x09, 0x20, 0x48, 0x29,
-    0x7f, 0x8d, 0x7b, 0x06, 0x68, 0x38, 0x60, 0x7b, 0x50, 0x5e, 0x29, 0x1b, 0x08, 0x18, 0x19, 0x00, 0x08, 0xe0, 0x08, 0x00, 0x00, 0x00,
-    0x00, 0x8d, 0x7b, 0x06, 0xa5, 0x25, 0xcd, 0xfb, 0x05, 0xf0, 0x06, 0x8d, 0xfb, 0x05, 0x20, 0x04, 0xca, 0xa5, 0x24, 0xcd, 0x7b, 0x05,
-    0x90, 0x03, 0x8d, 0x7b, 0x05, 0xad, 0x7b, 0x06, 0x20, 0x89, 0xca, 0xa9, 0x0f, 0x8d, 0xb0, 0xc0, 0xad, 0x7b, 0x05, 0xc9, 0x50, 0xb0,
-    0x13, 0x6d, 0x7b, 0x04, 0x8d, 0xb1, 0xc0, 0xa9, 0x0e, 0x8d, 0xb0, 0xc0, 0xa9, 0x00, 0x6d, 0xfb, 0x04, 0x8d, 0xb1, 0xc0, 0x60, 0x49,
-    0xc0, 0xc9, 0x08, 0xb0, 0x1d, 0xa8, 0xa9, 0xc9, 0x48, 0xb9, 0xf2, 0xcb, 0x48, 0x60, 0xea, 0xac, 0x7b, 0x05, 0xa9, 0xa0, 0x20, 0x71,
-    0xca, 0xc8, 0xc0, 0x50, 0x90, 0xf8, 0x60, 0xa9, 0x34, 0x8d, 0x7b, 0x07, 0x60, 0xa9, 0x32, 0xd0, 0xf8, 0xa0, 0xc0, 0xa2, 0x80, 0xca,
-    0xd0, 0xfd, 0xad, 0x30, 0xc0, 0x88, 0xd0, 0xf5, 0x60, 0xac, 0x7b, 0x05, 0xc0, 0x50, 0x90, 0x05, 0x48, 0x20, 0xb0, 0xc9, 0x68, 0xac,
-    0x7b, 0x05, 0x20, 0x71, 0xca, 0xee, 0x7b, 0x05, 0x2c, 0x78, 0x04, 0x10, 0x07, 0xad, 0x7b, 0x05, 0xc9, 0x50, 0xb0, 0x68, 0x60, 0xac,
-    0x7b, 0x05, 0xad, 0xfb, 0x05, 0x48, 0x20, 0x07, 0xca, 0x20, 0x04, 0xc9, 0xa0, 0x00, 0x68, 0x69, 0x00, 0xc9, 0x18, 0x90, 0xf0, 0xb0,
-    0x23, 0x20, 0x67, 0xc9, 0x98, 0xf0, 0xe8, 0xa9, 0x00, 0x8d, 0x7b, 0x05, 0x8d, 0xfb, 0x05, 0xa8, 0xf0, 0x12, 0xce, 0x7b, 0x05, 0x10,
-    0x9d, 0xa9, 0x4f, 0x8d, 0x7b, 0x05, 0xad, 0xfb, 0x05, 0xf0, 0x93, 0xce, 0xfb, 0x05, 0x4c, 0x04, 0xca, 0xa9, 0x30, 0x8d, 0x7b, 0x07,
-    0x68, 0x09, 0x80, 0xc9, 0xb1, 0xd0, 0x67, 0xa9, 0x08, 0x8d, 0x58, 0xc0, 0xd0, 0x5b, 0xc9, 0xb2, 0xd0, 0x51, 0xa9, 0xfe, 0x2d, 0xfb,
-    0x07, 0x8d, 0xfb, 0x07, 0x60, 0x8d, 0x7b, 0x06, 0x4e, 0x78, 0x04, 0x4c, 0xcb, 0xc8, 0x20, 0x27, 0xca, 0xee, 0xfb, 0x05, 0xad, 0xfb,
-    0x05, 0xc9, 0x18, 0x90, 0x4a, 0xce, 0xfb, 0x05, 0xad, 0xfb, 0x06, 0x69, 0x04, 0x29, 0x7f, 0x8d, 0xfb, 0x06, 0x20, 0x12, 0xca, 0xa9,
-    0x0d, 0x8d, 0xb0, 0xc0, 0xad, 0x7b, 0x04, 0x8d, 0xb1, 0xc0, 0xa9, 0x0c, 0x8d, 0xb0, 0xc0, 0xad, 0xfb, 0x04, 0x8d, 0xb1, 0xc0, 0xa9,
-    0x17, 0x20, 0x07, 0xca, 0xa0, 0x00, 0x20, 0x04, 0xc9, 0xb0, 0x95, 0xc9, 0xb3, 0xd0, 0x0e, 0xa9, 0x01, 0x0d, 0xfb, 0x07, 0xd0, 0xa9,
-    0xc9, 0xb0, 0xd0, 0x9c, 0x4c, 0x09, 0xc8, 0x4c, 0x27, 0xc9, 0xad, 0xfb, 0x05, 0x8d, 0xf8, 0x04, 0x0a, 0x0a, 0x6d, 0xf8, 0x04, 0x6d,
-    0xfb, 0x06, 0x48, 0x4a, 0x4a, 0x4a, 0x4a, 0x8d, 0xfb, 0x04, 0x68, 0x0a, 0x0a, 0x0a, 0x0a, 0x8d, 0x7b, 0x04, 0x60, 0xc9, 0x0d, 0xd0,
-    0x06, 0xa9, 0x00, 0x8d, 0x7b, 0x05, 0x60, 0x09, 0x80, 0xc9, 0xa0, 0xb0, 0xce, 0xc9, 0x87, 0x90, 0x08, 0xa8, 0xa9, 0xc9, 0x48, 0xb9,
-    0xb9, 0xc9, 0x48, 0x60, 0x18, 0x71, 0x13, 0xb2, 0x48, 0x60, 0xaf, 0x9d, 0xf2, 0x13, 0x13, 0x13, 0x13, 0x13, 0x13, 0x13, 0x13, 0x13,
-    0x66, 0x0e, 0x13, 0x38, 0x00, 0x14, 0x7b, 0x18, 0x98, 0x6d, 0x7b, 0x04, 0x48, 0xa9, 0x00, 0x6d, 0xfb, 0x04, 0x48, 0x0a, 0x29, 0x0c,
-    0xaa, 0xbd, 0xb0, 0xc0, 0x68, 0x4a, 0x68, 0xaa, 0x60, 0x0a, 0x48, 0xad, 0xfb, 0x07, 0x4a, 0x68, 0x6a, 0x48, 0x20, 0x59, 0xca, 0x68,
-    0xb0, 0x05, 0x9d, 0x00, 0xcc, 0x90, 0x03, 0x9d, 0x00, 0xcd, 0x60, 0x48, 0xa9, 0xf7, 0x20, 0xa0, 0xc9, 0x8d, 0x59, 0xc0, 0xad, 0x7b,
-    0x07, 0x29, 0x07, 0xd0, 0x04, 0x68, 0x4c, 0x23, 0xca, 0x29, 0x04, 0xf0, 0x03, 0x4c, 0x87, 0xc9, 0x68, 0x38, 0xe9, 0x20, 0x29, 0x7f,
-    0x48, 0xce, 0x7b, 0x07, 0xad, 0x7b, 0x07, 0x29, 0x03, 0xd0, 0x15, 0x68, 0xc9, 0x18, 0xb0, 0x03, 0x8d, 0xfb, 0x05, 0xad, 0xf8, 0x05,
-    0xc9, 0x50, 0xb0, 0x03, 0x8d, 0x7b, 0x05, 0x4c, 0x04, 0xca, 0x68, 0x8d, 0xf8, 0x05, 0x60, 0xad, 0x00, 0xc0, 0xc9, 0x93, 0xd0, 0x0f,
-    0x2c, 0x10, 0xc0, 0xad, 0x00, 0xc0, 0x10, 0xfb, 0xc9, 0x83, 0xf0, 0x03, 0x2c, 0x10, 0xc0, 0x60, 0xa8, 0xb9, 0x31, 0xcb, 0x20, 0xf1,
-    0xc8, 0x20, 0x44, 0xc8, 0xc9, 0xce, 0xb0, 0x08, 0xc9, 0xc9, 0x90, 0x04, 0xc9, 0xcc, 0xd0, 0xea, 0x4c, 0xf1, 0xc8, 0xea, 0x2c, 0xcb,
-    0xff, 0x70, 0x31, 0x38, 0x90, 0x18, 0xb8, 0x50, 0x2b, 0x01, 0x82, 0x11, 0x14, 0x1c, 0x22, 0x4c, 0x00, 0xc8, 0x20, 0x44, 0xc8, 0x29,
-    0x7f, 0xa2, 0x00, 0x60, 0x20, 0xa7, 0xc9, 0xa2, 0x00, 0x60, 0xc9, 0x00, 0xf0, 0x09, 0xad, 0x00, 0xc0, 0x0a, 0x90, 0x03, 0x20, 0x5c,
-    0xc8, 0xa2, 0x00, 0x60, 0x91, 0x28, 0x38, 0xb8, 0x8d, 0xff, 0xcf, 0x48, 0x85, 0x35, 0x8a, 0x48, 0x98, 0x48, 0xa5, 0x35, 0x86, 0x35,
-    0xa2, 0xc3, 0x8e, 0x78, 0x04, 0x48, 0x50, 0x10, 0xa9, 0x32, 0x85, 0x38, 0x86, 0x39, 0xa9, 0x07, 0x85, 0x36, 0x86, 0x37, 0x20, 0x00,
-    0xc8, 0x18, 0x90, 0x6f, 0x68, 0xa4, 0x35, 0xf0, 0x1f, 0x88, 0xad, 0x78, 0x06, 0xc9, 0x88, 0xf0, 0x17, 0xd9, 0x00, 0x02, 0xf0, 0x12,
-    0x49, 0x20, 0xd9, 0x00, 0x02, 0xd0, 0x3b, 0xad, 0x78, 0x06, 0x99, 0x00, 0x02, 0xb0, 0x03, 0x20, 0xed, 0xca, 0xa9, 0x80, 0x20, 0xf5,
-    0xc9, 0x20, 0x44, 0xc8, 0xc9, 0x9b, 0xf0, 0xf1, 0xc9, 0x8d, 0xd0, 0x05, 0x48, 0x20, 0x01, 0xc9, 0x68, 0xc9, 0x95, 0xd0, 0x12, 0xac,
-    0x7b, 0x05, 0x20, 0x59, 0xca, 0xb0, 0x05, 0xbd, 0x00, 0xcc, 0x90, 0x03, 0xbd, 0x00, 0xcd, 0x09, 0x80, 0x8d, 0x78, 0x06, 0xd0, 0x08,
-    0x20, 0x44, 0xc8, 0xa0, 0x00, 0x8c, 0x78, 0x06, 0xba, 0xe8, 0xe8, 0xe8, 0x9d, 0x00, 0x01, 0xa9, 0x00, 0x85, 0x24, 0xad, 0xfb, 0x05,
-    0x85, 0x25, 0x4c, 0x2e, 0xc8, 0x68, 0xac, 0xfb, 0x07, 0x10, 0x08, 0xac, 0x78, 0x06, 0xc0, 0xe0, 0x90, 0x01, 0x98, 0x20, 0xb1, 0xc8,
-    0x20, 0xcf, 0xca, 0xa9, 0x7f, 0x20, 0xa0, 0xc9, 0xad, 0x7b, 0x05, 0xe9, 0x47, 0x90, 0xd4, 0x69, 0x1f, 0x18, 0x90, 0xd1, 0x60, 0x38,
-    0x71, 0xb2, 0x7b, 0x00, 0x48, 0x66, 0xc4, 0xc2, 0xc1, 0xff, 0xc3, 0xea};
-uint8_t videx_chars[0xfff];  // 0x0-0x7ff rom 1 | 0x800-0xfff rom2 character rom (12x8) Matrix)
-uint8_t videx_vram[2048];    // 0x0800 at CC00-CDFF in 4 banks of 0x1FF
-
-// internal variables
-uint8_t videx_bankSLOT;
-uint8_t videx_regvalSLOT;                  // active memory bank, selected register
-uint8_t videx_regSLOT[17];                 // registers of the CRT-controller
-uint8_t videx_upperSLOT, videx_lowerSLOT;  // cursor size
-uint8_t videx_blink_mode, videx_blink_cursor;
-
-int card_videx_getC0SLOTX(int adr)  // 0x0400 at C800-CBFF
-{
-    int value = 0x00;
-
-    // define current memory bank
-    videx_bankSLOT = (adr & 0x000c) >> 2;
-
-    if(adr & 0x0001)
-        // get current register value
-        value = videx_regSLOT[videx_regvalSLOT];
-    else
-        // define current register
-        value = videx_regvalSLOT;
-
-    return value;
-}
-
-
-// Shadow accesses to card registers in $C0n0 - $C0nF range
-void videx_shadow_register(bool is_write, uint_fast16_t address, uint_fast8_t data) {
-    // define current memory bank
-    videx_bankSLOT = (address & 0x000c) >> 2;
-
-    if(!is_write)
-        return;
-
-    if(address & 0x0001) {
-        // set current register value
-        videx_regSLOT[videx_regvalSLOT] = data;
-
-        // 10: Cursor upper, 11: Cursor lower
-        if((videx_regvalSLOT == 10) || (videx_regvalSLOT == 11)) {
-            card_videx_modifySLOT();
-        }
-    } else {
-        // define current register
-        videx_regvalSLOT = data;
-    }
-}
-
-
-// IOSEL READS ROM
-// when reading using iosel, adds bit at A9
-uint8_t card_videx_getRomIoSel(int addr) {
-    if(addr < 0xCC00)
-        return videx_rom[(addr | 0b1000000000) & 0x03ff];
-
-    return 0;
-}
-
-// Read Video Memory using position x, y
-int8_t card_videx_getCSLOTXX(int col, int row) {
-    int vstart = ((videx_regSLOT[12] << 8) | videx_regSLOT[13]);
-    // int vcursor = ((videx_regSLOT[14] << 8) | videx_regSLOT[15]);
-
-    int vaddr = (((row * 80) + col) + vstart) % 0x800;
-
-    return videx_vram[vaddr];
-}
-
-// Read Character Rom (Char/scanline)
-int8_t card_videx_getVideoBits(int8_t keycode, int line) {
-    return videx_chars[keycode * 16 + line];
-}
-
-// Is cursor at defined x, y?
-bool card_videx_isCursorPosition(int col, int row) {
-    int vstart = ((videx_regSLOT[12] << 8) | videx_regSLOT[13]);
-    int vcursor = ((videx_regSLOT[14] << 8) | videx_regSLOT[15]);
-    return vcursor == (((row * 80) + col) + vstart);
-}
-
-// IOSTB - Get Rom or VRam
-int card_videx_getSLOTC8XX(int adr) {
-    // if (adr < 0x0400) // Rom is at 0x0-0x3ff
-    if(adr < 0xCC00)  // Rom is at 0x0-0x3ff
-        return (videx_rom[adr & 0x03ff]);
-    else
-        // VRAM is at 0x600-0x7FF
-        // VRAM has 0x0800 at CC00-CDFF divided in 4 banks of 0x1FF
-        // if (adr < 0x0600)
-        return (videx_vram[(adr & 0x01ff) + videx_bankSLOT * 0x0200]);
-    // else //600 (or CE00-CF00)
-    //	return(0);
-}
-
-// IOSTB - set VRam
-void card_videx_putSLOTC8XX(int adr, int value) {
-    // VRAM is at 0x600-0x7FF
-    // VRAM has 0x0800 at CC00-CDFF divided in 4 banks of 0x1FF
-    if((adr >= 0xCC00) && (adr <= 0xCDFF)) {
-        int vadr;
-
-        vadr = (adr & 0x01ff) + videx_bankSLOT * 0x0200;
-        videx_vram[vadr] = value;
-    }
-}
-
-
-void card_videx_card_videx_init() {
-    // initializing registers
-    videx_bankSLOT = 0;
-    videx_regvalSLOT = 0;
-    videx_regSLOT[0] = 0x7b;
-    videx_regSLOT[1] = 0x50;
-    videx_regSLOT[2] = 0x62;
-    videx_regSLOT[3] = 0x29;
-    videx_regSLOT[4] = 0x1b;
-    videx_regSLOT[5] = 0x08;
-    videx_regSLOT[6] = 0x18;
-    videx_regSLOT[7] = 0x19;
-    videx_regSLOT[8] = 0x0;
-    videx_regSLOT[9] = 0x8;
-    videx_regSLOT[10] = 0xc0;
-    videx_regSLOT[11] = 0x8;
-    videx_regSLOT[12] = 0x0;
-    videx_regSLOT[13] = 0x0;
-    videx_regSLOT[14] = 0x0;
-    videx_regSLOT[15] = 0x0;
-    videx_regSLOT[16] = 0x0;
-    videx_upperSLOT = 0;
-    videx_lowerSLOT = 8;
-    // videx_oldcursorSLOT = 0;
-
-    card_videx_modifySLOT();
-}
-
-void card_videx_modifySLOT() {
-    videx_blink_cursor = (videx_regSLOT[10] & 0x40) > 0;
-    videx_blink_mode = (videx_regSLOT[10] & 0x20) > 0;
-
-    // If videx_blink_cursor is false then
-    // videx_blink_mode determines if there is a cursor displayed (false) or not (true).
-
-    // If videx_blink_cursor is true then
-    // videx_blink_mode determines blink rate
-    //  - false = 1/16th field rate blink
-    //  - true = 1/32th field rate blink
-
-    // glyph start
-    videx_upperSLOT = videx_regSLOT[10] & 0xf;
-    if(videx_upperSLOT > 11)
-        videx_upperSLOT = 11;
-
-    // glyph end
-    videx_lowerSLOT = videx_regSLOT[11] & 0xf;
-    if(videx_lowerSLOT > 11)
-        videx_lowerSLOT = 11;
-
-    // if the setting is ridiculous
-    if(videx_upperSLOT >= videx_lowerSLOT) {
-        videx_upperSLOT = 0;
-        videx_lowerSLOT = 11;
-    }
-}
